backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
asyncpg==0.29.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
python-dateutil==2.8.2
scikit-learn==1.3.0
skl2onnx==1.16.0
openai>=1.0.0
requests==2.31.0
azure-ai-formrecognizer==3.3.0